Duplicated code removed from Receive in TcpL4Protocol
When TcpL4Protocol receive a packet, the methods which handle the management
of the packet itself (for both IPv4 and IPv6) shares a lot of code. Differencies
are only in the endpoint searching; the other operations have been merged in two
protected methods, which are called by the Receive methods.